refactor: pulled out interface TensorValidator
